support uint64

+++ internal/testdata/states.go
@@ -20,7 +20,7 @@ type outcomes struct {
 	// Create returns a simple outcome {a: aBalance, b: bBalance} with the supplied
 	// erc20 token address as the asset. A zero address implies the
 	// zero-asset (chain-native token)
-	Create func(a, b types.Address, aBalance, bBalance uint, token common.Address) outcome.Exit
+	Create func(a, b types.Address, aBalance, bBalance uint64, token common.Address) outcome.Exit
 	// CreateLongOutcome returns a simple outcome {addressOne: balanceOne ...} in the
 	// zero-asset (chain-native token)
 	// The outcome can be of arbitrary length, and is formed in order that the SimpleItems are provided
@@ -77,7 +77,7 @@ func createLedgerState(a, b types.Address, aBalance, bBalance uint) state.State 
 		a,
 		b,
 	}
-	state.Outcome = Outcomes.Create(a, b, aBalance, bBalance, common.Address{})
+	state.Outcome = Outcomes.Create(a, b, uint64(aBalance), uint64(bBalance), common.Address{})
 	state.AppDefinition = types.Address{} // ledger channel running the consensus app
 	state.TurnNum = 0
 
@@ -85,7 +85,7 @@ func createLedgerState(a, b types.Address, aBalance, bBalance uint) state.State 
 }
 
 // createOutcome is a helper function to create a two-actor outcome
-func createOutcome(first types.Address, second types.Address, x, y uint, asset common.Address) outcome.Exit {
+func createOutcome(first types.Address, second types.Address, x, y uint64, asset common.Address) outcome.Exit {
 	return outcome.Exit{outcome.SingleAssetExit{
 		Asset: asset,
 		Allocations: outcome.Allocations{
